# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Spatial Memory Network - an AR app that lets users pin multimedia memories to real-world geographic locations.",
        "title": "Spatial Memory Network API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/spatial-memory/spatial-memory/issues",
            "email": "support@spatial-memory.app"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/memories": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new memory at a geographic location",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memories"
                ],
                "summary": "Create a new memory",
                "parameters": [
                    {
                        "description": "Memory data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateMemoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created memory",
                        "schema": {
                            "$ref": "#/definitions/model.Memory"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/memories/mine": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a paginated list of the authenticated user's memories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memories"
                ],
                "summary": "List user's memories",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 20)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of memories",
                        "schema": {
                            "$ref": "#/definitions/response.PaginatedResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/memories/nearby": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Find memories near a geographic location",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memories"
                ],
                "summary": "Find nearby memories",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Latitude (-90 to 90)",
                        "name": "lat",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Longitude (-180 to 180)",
                        "name": "lng",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Search radius in meters (default: 1000, max: 50000)",
                        "name": "radius",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by: distance, recent, popular (default: distance)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum results (default: 20, max: 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Enable clustering for large result sets",
                        "name": "cluster",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of nearby memories",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Memory"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/memories/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a memory's details by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memories"
                ],
                "summary": "Get a memory by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Memory ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Memory details",
                        "schema": {
                            "$ref": "#/definitions/model.Memory"
                        }
                    },
                    "400": {
                        "description": "Invalid memory ID",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Memory not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update a memory's title, content, or visibility",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memories"
                ],
                "summary": "Update a memory",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Memory ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update fields",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateMemoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated memory",
                        "schema": {
                            "$ref": "#/definitions/model.Memory"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Memory not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Soft delete a memory by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memories"
                ],
                "summary": "Delete a memory",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Memory ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid memory ID",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CreateMemoryRequest": {
            "type": "object",
            "required": [
                "location",
                "title"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "maxLength": 500
                },
                "content": {
                    "type": "string",
                    "maxLength": 5000
                },
                "location": {
                    "$ref": "#/definitions/model.GeoPoint"
                },
                "title": {
                    "type": "string",
                    "maxLength": 200
                },
                "visibility": {
                    "maximum": 2,
                    "minimum": 0,
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Visibility"
                        }
                    ]
                }
            }
        },
...
        "model.GeoPoint": {
            "type": "object",
            "required": [
                "lat",
                "lng"
            ],
            "properties": {
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                }
            }
        },
...
        "model.MediaType": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-varnames": [
                "MediaTypePhoto",
                "MediaTypeVideo",
                "MediaTypeVoice"
            ]
        },
        "model.Memory": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "bookmark_count": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "distance_meters": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "is_bookmarked": {
                    "type": "boolean"
                },
                "is_liked": {
                    "type": "boolean"
                },
                "like_count": {
                    "type": "integer"
                },
                "location": {
                    "$ref": "#/definitions/model.GeoPoint"
                },
                "media": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.MemoryMedia"
                    }
                },
                "status": {
                    "$ref": "#/definitions/model.MemoryStatus"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "view_count": {
                    "type": "integer"
                },
                "visibility": {
                    "$ref": "#/definitions/model.Visibility"
                }
            }
        },
        "model.MemoryMedia": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "duration": {
                    "description": "Alias for DurationSeconds",
                    "type": "integer"
                },
                "duration_seconds": {
                    "description": "DB field",
                    "type": "integer"
                },
                "file_size": {
                    "description": "Alias for SizeBytes",
                    "type": "integer"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "media_type": {
                    "$ref": "#/definitions/model.MediaType"
                },
                "memory_id": {
                    "type": "integer"
                },
                "mime_type": {
                    "type": "string"
                },
                "size_bytes": {
                    "description": "DB field",
                    "type": "integer"
                },
                "sort_order": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "model.MemoryStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "MemoryStatusDeleted",
                "MemoryStatusActive",
                "MemoryStatusPendingReview",
                "MemoryStatusRejected"
            ]
        },
...
        "model.UpdateMemoryRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 5000
                },
                "title": {
                    "type": "string",
                    "maxLength": 200
                },
                "visibility": {
                    "maximum": 2,
                    "minimum": 0,
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Visibility"
                        }
                    ]
                }
            }
        },
...
        "model.Visibility": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-varnames": [
                "VisibilityPrivate",
                "VisibilityCircle",
                "VisibilityPublic"
            ]
        },
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 40000
                },
                "message": {
                    "type": "string",
                    "example": "invalid request"
                }
            }
        },
        "response.PaginatedData": {
            "type": "object",
            "properties": {
                "items": {},
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 20
                },
                "total": {
                    "type": "integer",
                    "example": 100
                },
                "total_pages": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "response.PaginatedResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/response.PaginatedData"
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}